Decouple RPC from ABCI domain types.
The original contents of `tendermint::abci` were created only to model RPC
responses, not to model ABCI itself. The old types are shaped to fit the
details of the JSON encoding used by the RPC. Eventually, they should be
eliminated and merged with the new ABCI domain types. Moving them to the RPC
crate in the meantime disentangles that work from the work of finishing the
modeling of the ABCI domain types.